requirements/requirements.txt
requests==2.31.0
python-chess==1.999